+++ .idea/dictionaries/Phoenixo.xml
@@ -1,0 +1,11 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="Phoenixo">
+    <words>
+      <w>fclose</w>
+      <w>fopen</w>
+      <w>fputs</w>
+      <w>printf</w>
+      <w>scanf</w>
+    </words>
+  </dictionary>
+</component>
