+++ cmake-build-debug/CMakeFiles/Makefile2
@@ -66,7 +66,7 @@ CMAKE_BINARY_DIR = /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/cmake-
 CMakeFiles/msl_clang_001.dir/all:
 	$(MAKE) -f CMakeFiles/msl_clang_001.dir/build.make CMakeFiles/msl_clang_001.dir/depend
 	$(MAKE) -f CMakeFiles/msl_clang_001.dir/build.make CMakeFiles/msl_clang_001.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/cmake-build-debug/CMakeFiles --progress-num=1,2 "Built target msl_clang_001"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=/Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/cmake-build-debug/CMakeFiles --progress-num=1,2,3 "Built target msl_clang_001"
 .PHONY : CMakeFiles/msl_clang_001.dir/all
 
 # Include target in all.
@@ -76,7 +76,7 @@ all: CMakeFiles/msl_clang_001.dir/all
 
 # Build rule for subdir invocation for target.
 CMakeFiles/msl_clang_001.dir/rule: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/cmake-build-debug/CMakeFiles 2
+	$(CMAKE_COMMAND) -E cmake_progress_start /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/cmake-build-debug/CMakeFiles 3
 	$(MAKE) -f CMakeFiles/Makefile2 CMakeFiles/msl_clang_001.dir/all
 	$(CMAKE_COMMAND) -E cmake_progress_start /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/cmake-build-debug/CMakeFiles 0
 .PHONY : CMakeFiles/msl_clang_001.dir/rule
