+++ cmake-build-debug/CMakeFiles/msl_clang_001.dir/build.make
@@ -81,17 +81,43 @@ CMakeFiles/msl_clang_001.dir/main.c.o.provides: CMakeFiles/msl_clang_001.dir/mai
 CMakeFiles/msl_clang_001.dir/main.c.o.provides.build: CMakeFiles/msl_clang_001.dir/main.c.o
 
 
+CMakeFiles/msl_clang_001.dir/BSTree.c.o: CMakeFiles/msl_clang_001.dir/flags.make
+CMakeFiles/msl_clang_001.dir/BSTree.c.o: ../BSTree.c
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --progress-dir=/Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_2) "Building C object CMakeFiles/msl_clang_001.dir/BSTree.c.o"
+	/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -o CMakeFiles/msl_clang_001.dir/BSTree.c.o   -c /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/BSTree.c
+
+CMakeFiles/msl_clang_001.dir/BSTree.c.i: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Preprocessing C source to CMakeFiles/msl_clang_001.dir/BSTree.c.i"
+	/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -E /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/BSTree.c > CMakeFiles/msl_clang_001.dir/BSTree.c.i
+
+CMakeFiles/msl_clang_001.dir/BSTree.c.s: cmake_force
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green "Compiling C source to assembly CMakeFiles/msl_clang_001.dir/BSTree.c.s"
+	/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/cc $(C_DEFINES) $(C_INCLUDES) $(C_FLAGS) -S /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/BSTree.c -o CMakeFiles/msl_clang_001.dir/BSTree.c.s
+
+CMakeFiles/msl_clang_001.dir/BSTree.c.o.requires:
+
+.PHONY : CMakeFiles/msl_clang_001.dir/BSTree.c.o.requires
+
+CMakeFiles/msl_clang_001.dir/BSTree.c.o.provides: CMakeFiles/msl_clang_001.dir/BSTree.c.o.requires
+	$(MAKE) -f CMakeFiles/msl_clang_001.dir/build.make CMakeFiles/msl_clang_001.dir/BSTree.c.o.provides.build
+.PHONY : CMakeFiles/msl_clang_001.dir/BSTree.c.o.provides
+
+CMakeFiles/msl_clang_001.dir/BSTree.c.o.provides.build: CMakeFiles/msl_clang_001.dir/BSTree.c.o
+
+
 # Object files for target msl_clang_001
 msl_clang_001_OBJECTS = \
-"CMakeFiles/msl_clang_001.dir/main.c.o"
+"CMakeFiles/msl_clang_001.dir/main.c.o" \
+"CMakeFiles/msl_clang_001.dir/BSTree.c.o"
 
 # External object files for target msl_clang_001
 msl_clang_001_EXTERNAL_OBJECTS =
 
 msl_clang_001: CMakeFiles/msl_clang_001.dir/main.c.o
+msl_clang_001: CMakeFiles/msl_clang_001.dir/BSTree.c.o
 msl_clang_001: CMakeFiles/msl_clang_001.dir/build.make
 msl_clang_001: CMakeFiles/msl_clang_001.dir/link.txt
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_2) "Linking C executable msl_clang_001"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --green --bold --progress-dir=/Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/cmake-build-debug/CMakeFiles --progress-num=$(CMAKE_PROGRESS_3) "Linking C executable msl_clang_001"
 	$(CMAKE_COMMAND) -E cmake_link_script CMakeFiles/msl_clang_001.dir/link.txt --verbose=$(VERBOSE)
 
 # Rule to build all files generated by this target.
@@ -100,6 +126,7 @@ CMakeFiles/msl_clang_001.dir/build: msl_clang_001
 .PHONY : CMakeFiles/msl_clang_001.dir/build
 
 CMakeFiles/msl_clang_001.dir/requires: CMakeFiles/msl_clang_001.dir/main.c.o.requires
+CMakeFiles/msl_clang_001.dir/requires: CMakeFiles/msl_clang_001.dir/BSTree.c.o.requires
 
 .PHONY : CMakeFiles/msl_clang_001.dir/requires
