+++ cmake-build-debug/CMakeFiles/msl_clang_001.dir/depend.internal
@@ -1,5 +1,9 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 3.9
 
+CMakeFiles/msl_clang_001.dir/BSTree.c.o
+ /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/BSTree.c
+ /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/BSTree.h
 CMakeFiles/msl_clang_001.dir/main.c.o
+ /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/BSTree.h
  /Users/Phoenixo/Documents/Projects/CLion/msl-clang-001/main.c
