+++ cmake-build-debug/CMakeFiles/msl_clang_001.dir/depend.make
@@ -1,5 +1,9 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 3.9
 
+CMakeFiles/msl_clang_001.dir/BSTree.c.o: ../BSTree.c
+CMakeFiles/msl_clang_001.dir/BSTree.c.o: ../BSTree.h
+
+CMakeFiles/msl_clang_001.dir/main.c.o: ../BSTree.h
 CMakeFiles/msl_clang_001.dir/main.c.o: ../main.c
